requirements.txt
pandas
numpy
matplotlib
notebook
PyYAML
seaborn
scikit-learn
flask
boto3
lightgbm==4.0.0
xgboost==1.7.6
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2

